Added Restart feature
This was very difficult to do. I wanted to do this since refreshing the page would interfere with the Spotify playlist. There were issues of characters not being able to be able to move after the game had restarted or their death animation playing infinitely while  being able to move. However, it has now been resolved.