Added Power Feature
Fighter now have their own power feature which determines how much damage they inflict on others. This is done to try and balance the characters since their attack frames are different. Their was an issue where the medieval characters' health bar wouldn't fully go down or it would go back up after dying. This was due to the health being below 0 and the width having to adjust to a negative number. This has been resolved.