Updated Spotify Playlist Image
I added some new songs in the Spotify playlist and added a new image to the thumbnail of the playlist. I also adjusted some values for the boundaries to ensure that when both players are all the way to one side of the screen they're still able to hit each other. Otherwise it's an easy way to win.